implemented page forwarding, prevent from accessing jsp files directly
