+++ src/service/FavorService.java
@@ -4,4 +4,6 @@ import entity.FavorEntity;
 
 public interface FavorService {
     void insert(FavorEntity favor);
+
+    void delete(String favorId);
 }
