+++ src/servlet/AdminServlet.java
@@ -13,10 +13,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-
-import java.io.IOException;
 import java.io.*;
-
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -38,9 +35,7 @@ public class AdminServlet extends HttpServlet {
     @SuppressWarnings("unused")
     private void index(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String tsStr = "2011-05-09 11:49:45";
-
-        UsersEntity usersEntity = new UsersEntity("zhsyy","123@email","123456","admin", null);
-
+        UsersEntity usersEntity = new UsersEntity("zhsyy","123@email","123456","admin","");
         usersEntity.setTime(Timestamp.valueOf(tsStr));
         usersEntity.setUserId(2);
         HttpSession session = req.getSession();
@@ -52,7 +47,6 @@ public class AdminServlet extends HttpServlet {
     private void user(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String name = req.getParameter("name");
         UsersEntity usersEntity = userService.getUserByName(name);
-
         req.setAttribute("user",usersEntity);
         req.getRequestDispatcher("adminUser.jsp").forward(req, resp);
     }
@@ -61,20 +55,12 @@ public class AdminServlet extends HttpServlet {
     private void usersList(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         List<UsersEntity> list = userService.getAllUsers();
         req.setAttribute("usersList",list);
-
-//        System.out.println(list.size());
-
         req.getRequestDispatcher("adminUsersList.jsp").forward(req, resp);
     }
 
     @SuppressWarnings("unused")
     private void artwork(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String artworkId = req.getParameter("artworkId");
-
-        ArtworksEntity artworksEntity = artworkService.getArtwork(Integer.parseInt(artworkId));
-        req.setAttribute("artwork",artworksEntity);
-        req.getRequestDispatcher("adminUser.jsp").forward(req, resp);
-
         if (artworkId!=null){
             ArtworksEntity artworksEntity = artworkService.getArtwork(Integer.parseInt(artworkId));
             req.setAttribute("artwork",artworksEntity);
@@ -116,7 +102,7 @@ public class AdminServlet extends HttpServlet {
         String Email = req.getParameter("email");
         String Password = req.getParameter("password");
         String Type = req.getParameter("type");
-        UsersEntity usersEntity = new UsersEntity(Username,Email,Password,Type);
+        UsersEntity usersEntity = new UsersEntity(Username,Email,Password,Type,"");
         userService.addUser(usersEntity);
         resp.setHeader("refresh", "0.1;url=usersList.admin");
     }
