--- web/arts.html
@@ -1,173 +1,0 @@
-<!doctype html>
-<html lang="en"><head>
-    <meta charset="utf-8">
-    <title>Bootstrap Admin</title>
-    <meta content="IE=edge,chrome=1" http-equiv="X-UA-Compatible">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="">
-    <meta name="author" content="">
-    <link href='http://fonts.googleapis.com/css?family=Open+Sans:400,700' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" type="text/css" href="lib/bootstrap/css/bootstrap.css">
-    <link rel="stylesheet" href="lib/font-awesome/css/font-awesome.css">
-    <script src="lib/jquery-1.11.1.min.js" type="text/javascript"></script>
-    <link rel="stylesheet" type="text/css" href="css/theme.css">
-    <link rel="stylesheet" type="text/css" href="css/premium.css">
-
-</head>
-<body class=" theme-blue">
-<style type="text/css">
-    .navbar-default .navbar-brand, .navbar-default .navbar-brand:hover {
-        color: #fff;
-    }
-</style>
-
-<script type="text/javascript">
-    $(function() {
-        var uls = $('.sidebar-nav > ul > *').clone();
-        uls.addClass('visible-xs');
-        $('#main-menu').append(uls.clone());
-    });
-</script>
-
-<div class="navbar navbar-default" role="navigation">
-    <div class="navbar-header">
-        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target=".navbar-collapse">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-        </button>
-        <a class="" href="adminIndex.jsp"><span class="navbar-brand"><span class="fa fa-paper-plane"></span> Aircraft</span></a></div>
-
-    <div class="navbar-collapse collapse" style="height: 1px;">
-        <ul id="main-menu" class="nav navbar-nav navbar-right">
-            <li class="dropdown hidden-xs">
-                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-                    <span class="glyphicon glyphicon-user padding-right-small" style="position:relative;top: 3px;"></span> Jack Smith
-                    <i class="fa fa-caret-down"></i>
-                </a>
-
-                <ul class="dropdown-menu">
-                    <li><a href="">My Account</a></li>
-                    <li class="divider"></li>
-                    <li class="dropdown-header">Admin Panel</li><!--这个好像没啥用，可以做装饰-->
-                    <li><a href="">Users</a></li>
-                    <li class="divider"></li>
-                    <li><a tabindex="-1" href="">Logout</a></li>
-                </ul>
-            </li>
-        </ul>
-    </div>
-</div>
-</div>
-
-
-<div class="sidebar-nav">
-    <ul>
-        <li><a href="#" data-target=".dashboard-menu" class="nav-header collapsed" data-toggle="collapse"><i class="fa fa-fw fa-dashboard"></i> Dashboard<i class="fa fa-collapse"></i></a></li>
-        <li><ul class="dashboard-menu nav nav-list collapse">
-            <li><a href="adminUser.jsp"><span class="fa fa-caret-right"></span> My Profile</a></li>
-            <li ><a href="adminUsersList.jsp"><span class="fa fa-caret-right"></span> User List</a></li>
-        </ul></li>
-
-        <li><a href="#" data-target=".legal-menu" class="nav-header" data-toggle="collapse"><i class="fa fa-fw fa-legal"></i> Legal<i class="fa fa-collapse"></i></a></li>
-        <li><ul class="legal-menu nav nav-list collapse in">
-            <li ><a href="adminArtwork.jsp"><span class="fa fa-caret-right"></span> Upload arts</a></li>
-            <li ><a href="arts.html"><span class="fa fa-caret-right"></span> Art list</a></li>
-        </ul></li>
-
-    </ul>
-</div>
-
-<div class="content">
-    <div class="header">
-
-        <h1 class="page-title">Arts</h1>
-
-    </div>
-    <div class="main-content">
-
-        <div class="btn-toolbar list-toolbar">
-            <button class="btn btn-primary" onclick="window.location.href='art.html'"><i class="fa fa-plus"></i>New Art</button><!--添加新的用户，可设置跳转到注册界面,不是一般用户的注册界面-->
-            <div class="btn-group">
-            </div>
-        </div>
-        <table class="table">
-            <thead>
-            <tr>
-                <th>ID</th>
-                <th>Name</th>
-                <th>Introduction</th>
-                <th>Picture</th>
-                <th>Location</th>
-                <th>Years</th>
-                <th style="width: 3.5em;"></th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <td>1</td>
-                <td>书</td>
-                <td>abc</td>
-                <!--此处信息可由url传入-->
-                <td><img style="width: 100px" src="xxx"></td>
-                <td>张江</td>
-                <td>2019-07-14 PM 10:00</td>
-                <td>
-                    <a href="adminArtwork.jsp"><i class="fa fa-pencil"></i></a><!--修改-->
-                    <a href="#myModal" role="button" data-toggle="modal"><i class="fa fa-trash-o"></i></a><!--删除-->
-                </td>
-            </tr>
-            </tbody>
-        </table>
-
-        <!--分页-->
-        <ul class="pagination">
-            <li><a href="#">&laquo;</a></li>
-            <li><a href="#">1</a></li>
-            <li><a href="#">2</a></li>
-            <li><a href="#">3</a></li>
-            <li><a href="#">4</a></li>
-            <li><a href="#">5</a></li>
-            <li><a href="#">&raquo;</a></li>
-        </ul>
-
-        <div class="modal small fade" id="myModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-            <div class="modal-dialog">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">×</button>
-                        <h3 id="myModalLabel">Delete Confirmation</h3>
-                    </div>
-                    <div class="modal-body">
-                        <p class="error-text"><i class="fa fa-warning modal-icon"></i>Are you sure you want to delete the art?<br>This cannot be undone.</p>
-                    </div>
-                    <div class="modal-footer">
-                        <button class="btn btn-default" data-dismiss="modal" aria-hidden="true">Cancel</button>
-                        <button class="btn btn-danger" data-dismiss="modal">Delete</button>
-                        <!--点击此处删除后从数据库删除，并刷新该页面-->
-                    </div>
-                </div>
-            </div>
-        </div>
-
-
-        <footer>
-            <hr>
-            <p class="pull-right">A <a href="./" target="_blank">Free Bootstrap Theme</a> by <a href="./" target="_blank">Portnine</a></p>
-            <p>© 2014 <a href="http://www.portnine.com" target="_blank">Portnine</a></p>
-        </footer>
-    </div>
-</div>
-
-
-<script src="lib/bootstrap/js/bootstrap.js"></script>
-<script type="text/javascript">
-    $("[rel=tooltip]").tooltip();
-    $(function() {
-        $('.demo-cancel-click').click(function(){return false;});
-    });
-</script>
-
-
-</body></html>
