--- web/sign-up.html
@@ -1,157 +1,0 @@
-<!doctype html>
-<html lang="en"><head>
-    <meta charset="utf-8">
-    <title>Bootstrap Admin</title>
-    <meta content="IE=edge,chrome=1" http-equiv="X-UA-Compatible">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="">
-    <meta name="author" content="">
-
-    <link href='http://fonts.googleapis.com/css?family=Open+Sans:400,700' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" type="text/css" href="lib/bootstrap/css/bootstrap.css">
-    <link rel="stylesheet" href="lib/font-awesome/css/font-awesome.css">
-
-    <script src="lib/jquery-1.11.1.min.js" type="text/javascript"></script>
-
-
-
-    <link rel="stylesheet" type="text/css" href="css/theme.css">
-    <link rel="stylesheet" type="text/css" href="css/premium.css">
-
-</head>
-<body class=" theme-blue">
-<script type="text/javascript">
-    $(function() {
-        var match = document.cookie.match(new RegExp('color=([^;]+)'));
-        if(match) var color = match[1];
-        if(color) {
-            $('body').removeClass(function (index, css) {
-                return (css.match (/\btheme-\S+/g) || []).join(' ')
-            })
-            $('body').addClass('theme-' + color);
-        }
-
-        $('[data-popover="true"]').popover({html: true});
-
-    });
-</script>
-<style type="text/css">
-    .navbar-default .navbar-brand, .navbar-default .navbar-brand:hover {
-        color: #fff;
-    }
-</style>
-
-<script type="text/javascript">
-    $(function() {
-        var uls = $('.sidebar-nav > ul > *').clone();
-        uls.addClass('visible-xs');
-        $('#main-menu').append(uls.clone());
-    });
-</script>
-<link rel="shortcut icon" href="../assets/ico/favicon.ico">
-<link rel="apple-touch-icon-precomposed" sizes="144x144" href="../assets/ico/apple-touch-icon-144-precomposed.png">
-<link rel="apple-touch-icon-precomposed" sizes="114x114" href="../assets/ico/apple-touch-icon-114-precomposed.png">
-<link rel="apple-touch-icon-precomposed" sizes="72x72" href="../assets/ico/apple-touch-icon-72-precomposed.png">
-<link rel="apple-touch-icon-precomposed" href="../assets/ico/apple-touch-icon-57-precomposed.png">
-<div class="navbar navbar-default" role="navigation">
-    <div class="navbar-header">
-        <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target=".navbar-collapse">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-        </button>
-        <a class="" href="adminIndex.jsp"><span class="navbar-brand"><span class="fa fa-paper-plane"></span> Aircraft</span></a></div>
-
-    <div class="navbar-collapse collapse" style="height: 1px;">
-        <ul id="main-menu" class="nav navbar-nav navbar-right">
-            <li class="dropdown hidden-xs">
-                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-                    <span class="glyphicon glyphicon-user padding-right-small" style="position:relative;top: 3px;"></span> Jack Smith
-                    <i class="fa fa-caret-down"></i>
-                </a>
-                <ul class="dropdown-menu">
-                    <li><a href="">My Account</a></li>
-                    <li class="divider"></li>
-                    <li class="dropdown-header">Admin Panel</li><!--这个好像没啥用，可以做装饰-->
-                    <li><a href="">Users</a></li>
-                    <li class="divider"></li>
-                    <li><a tabindex="-1" href="">Logout</a></li>
-                </ul>
-            </li>
-        </ul>
-
-    </div>
-</div>
-</div>
-
-
-<div class="sidebar-nav">
-    <ul>
-        <li><a href="#" data-target=".dashboard-menu" class="nav-header" data-toggle="collapse"><i class="fa fa-fw fa-dashboard"></i> Dashboard<i class="fa fa-collapse"></i></a></li>
-        <li><ul class="dashboard-menu nav nav-list collapse in">
-            <li><a href="adminUser.jsp"><span class="fa fa-caret-right"></span> My Profile</a></li>
-            <li ><a href="adminUsersList.jsp"><span class="fa fa-caret-right"></span> User List</a></li>
-        </ul></li>
-
-        <li><a href="#" data-target=".legal-menu" class="nav-header collapsed" data-toggle="collapse"><i class="fa fa-fw fa-legal"></i> Legal<i class="fa fa-collapse"></i></a></li>
-        <li><ul class="legal-menu nav nav-list collapse">
-            <li ><a href="adminArtwork.jsp"><span class="fa fa-caret-right"></span> Upload arts</a></li>
-            <li ><a href="arts.html"><span class="fa fa-caret-right"></span> Art list</a></li>
-        </ul></li>
-
-    </ul>
-</div>
-
-<div class="content">
-    <div class="header">
-        <h1 class="page-title">New User</h1>
-    </div>
-    <div class="main-content">
-
-        <div class="row">
-            <div class="col-md-4">
-                <br>
-                <div id="myTabContent" class="tab-content">
-                    <div class="tab-pane active in" id="home">
-                        <form id="tab">
-                            <div class="form-group">
-                                <label>Username</label>
-                                <input type="text" value="jsmith" class="form-control">
-                            </div>
-                            <div class="form-group">
-                                <label>Password</label>
-                                <input type="text" value="John" class="form-control">
-                            </div>
-                            <div class="form-group">
-                                <label>Email</label>
-                                <input type="text" value="Smith" class="form-control">
-                            </div>
-                        </form>
-                        <div class="btn-toolbar list-toolbar">
-                            <button class="btn btn-primary"><i class="fa fa-save"></i> Add</button>
-                            <button onclick="window.location.href='users.html'" data-toggle="modal" class="btn btn-danger">Back</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <footer>
-            <hr>
-            <p class="pull-right">A <a href="./" target="_blank">Free Bootstrap Theme</a> by <a href="./" target="_blank">Portnine</a></p>
-            <p>© 2014 <a href="http://www.portnine.com" target="_blank">Portnine</a></p>
-        </footer>
-    </div>
-</div>
-
-<script src="lib/bootstrap/js/bootstrap.js"></script>
-<script type="text/javascript">
-    $("[rel=tooltip]").tooltip();
-    $(function() {
-        $('.demo-cancel-click').click(function(){return false;});
-    });
-</script>
-
-
-</body></html>
